+++ src/Microsoft.Agents.A365.Observability.Contracts/Tracing/Scopes/AutoInstrumentationConstants.cs
@@ -6,7 +6,7 @@ namespace Microsoft.Agents.A365.Observability.Runtime.Tracing.Scopes
     /// <summary>
     /// Constants used for auto-instrumentation.
     /// </summary>
-    internal static class AutoInstrumentationConstants
+    public static class AutoInstrumentationConstants
     {
         /// <summary> The key for the input to a GenAI agent invocation. </summary>
         /// <remarks> Set by the Semantic Kernel OpenTelemetry integration for agent invocations.</remarks>